requirements.txt

Flask==2.0.1
line-bot-sdk==1.19.0
gunicorn==20.1.0
werkzeug==2.0.1
python-dotenv